Add Sentinel Mode
This commit makes the following changes to zdns

1. Adds a new sentinel module
The extensive module does a NS query followed by a A, AAAA, and SOA.

2. Adds a NSQ Streamhandler for Input and Output
Updates streamhandler.go to add ability to stream domain inputs
from NSQStreamhandler

3. JSONInput Format
Ability to take input in the form of JSON input. Looks for domain
and metadata.

4. Pass JSON Metadata
Code now passes along JSON metadata to the results as opposed to just a
string. The metadata can be used to embed scan_after time to enable
delayed scanning.